Add default path fallback
