+++ alvr/dashboard/src/steamvr_launcher/windows_steamvr.rs
@@ -1,8 +1,5 @@
 use std::os::windows::process::CommandExt;
-use std::path::PathBuf;
 use std::process::Command;
-
-use crate::steamvr_launcher::get_steamvr_root_dir;
 
 const CREATE_NO_WINDOW: u32 = 0x0800_0000;
 
@@ -20,14 +17,4 @@ pub fn kill_process(pid: u32) {
         .creation_flags(CREATE_NO_WINDOW)
         .output()
         .ok();
-}
-
-pub fn get_default_steamvr_executable_path() -> String {
-    return get_steamvr_root_dir()
-        .join("bin")
-        .join("win64")
-        .join("vrstartup.exe")
-        .into_os_string()
-        .into_string()
-        .unwrap();
 }
